make inline-editable a proper component
